# Add 'instance UniformRange Natural'