The main addition here is to add `instance UniformRange Natural`, also see https://github.com/haskell/random/issues/44.

To do this, I changed the `Integer`-specific `uniformIntegerM` to a generic `uniformIntegralM`. Since this function was and is recursive, I changed the pragma from `INLINE` to `INLINEABLE`.

I made sure that these more generic functions are no slower than before. Note that `random/Integer` benchmarks small ranges, for which a fast path is triggered; `uniformR/unbounded/Integer` and the new `uniformR/unbounded/Natural` benchmark very large ranges where there is no fast path.

Before (9ee79a7):
```
pure/random/Integer                      mean 362.6 μs  ( +- 44.38 μs  )
pure/uniformR/unbounded/Integer          mean 189.3 ms  ( +- 8.531 ms  )
```

After (this PR):
```
pure/random/Integer                      mean 369.0 μs  ( +- 36.89 μs  )
pure/uniformR/unbounded/Integer          mean 122.8 ms  ( +- 6.044 ms  )
pure/uniformR/unbounded/Natural          mean 110.6 ms  ( +- 16.08 ms  )
```

Other changes:
- specs use `uniformR` instead of `randomR`; this allow us to make the constraints more precise
- added `Natural` benchmark and specs
- sorted some imports alphabetically